Finished RecipeListAdapter
The display is correct for now.

Added another recipe to double check that if you had the ingredients in the Pantry but not enough of the ingredients in the Pantry it would say that you do not have all of the ingredients.

Later I will need to change the query to check against only the pantry for the logged in user.